Split ip-packet-requests types into modules
